### 2022-07-20 ff4473f
Use white Python logo
