+++ doc/source/index.md
@@ -5,11 +5,10 @@
    Tutorial <tutorial/index>
    Examples <examples/index>
    About <about>
-   GitHub <https://github.com/cbrnr/mnelab>
 ```
 
-![MNELAB GUI example](https://raw.githubusercontent.com/cbrnr/mnelab/main/mnelab.png)
+![MNELAB screenshots](https://raw.githubusercontent.com/cbrnr/mnelab/main/mnelab.png)
 
 MNELAB is a graphical user interface (GUI) for [MNE-Python](https://github.com/mne-tools/mne-python), a package for EEG/MEG analysis.
 
-Read the [**tutorial**](./tutorial/index) for installation instructions and first steps. The [**examples**](./examples/index) showcase MNELAB in common use cases.
+Read the [**tutorial**](./tutorial/index) for installation instructions and first steps. The [**examples**](./examples/index) showcase MNELAB for common use cases.
