+++ .github/workflows/release.yml
@@ -12,6 +12,8 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - uses: actions/setup-python@v4
+      with:
+        python-version: '3.10'
     - uses: pre-commit/action@v3.0.0
 
   build-wheel:
@@ -20,6 +22,8 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - uses: actions/setup-python@v4
+      with:
+        python-version: '3.10'
     - name: Build wheel
       run: |
         pip install wheel
@@ -35,7 +39,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
         with:
-          python-version: "3.8"
+          python-version: '3.10'
       - name: Build sdist
         run: |
           pip install build
