### 2022-08-04 #4595 @Kevin-Mattheus-Moerman
@brainstorm, @rougier, @satra, @tuliofalmeida, @effigies, @szorowi1, @agricolab, @sappelhoff, @ajquinn, @vahid-sb, @cbrnr, @DominiqueMakowski, @mgbleichner would you be interested in reviewing this work featuring a Python library for EEG for [JOSS](https://joss.theoj.org/).
JOSS reviews focus on [the software](https://github.com/cbrnr/mnelab) as well as [a short paper](https://github.com/openjournals/joss-papers/blob/joss.04595/joss.04595/10.21105.joss.04595.pdf). The review process takes place on GitHub. This here is a pre-review issue, where reviewers are assigned. Once we have enough reviewers I'll trigger a dedicated review issue, here is an example of an ongoing review: https://github.com/openjournals/joss-reviews/issues/4478.

If interested please comment here, thanks!

@AJQuinn, @vahid-sb, @tuliofalmeida, @anshu-97 of course you are already assigned as reviewers for https://github.com/openjournals/joss-reviews/issues/4478, let me know if you are also interested in reviewing this work. Thanks.
