### 2022-08-04 #4595 @vahid-sb
@Kevin-Mattheus-Moerman These days I  was actually thinking around the same need for a GUI myself. So, would be happy to review this library amd see to what extent it can help.
