### 2022-08-05 #4650 @editorialbot
[REVIEW]: MNELAB: a graphical user interface for MNE-Python
**Submitting author:** @cbrnr (<a href="http://orcid.org/0000-0002-6030-2233">Clemens Brunner</a>)
**Repository:** https://github.com/cbrnr/mnelab
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.8.5
