### 2022-08-05 #4650 @sappelhoff
@Kevin-Mattheus-Moerman reading the [conflict of interest guidelines](https://github.com/openjournals/joss/blob/main/COI.md) I realize that I do have a paper with @cbrnr from 2019 (also [published in JOSS](https://joss.theoj.org/papers/10.21105/joss.01896)). Furthermore, we both contribute to several repositories together, such as [mne-python](https://github.com/mne-tools/mne-python/) or [pybv](https://github.com/bids-standard/pybv).

However, I am sure that this will **not** impact my ability to make an impartial scientific judgment/evaluation of the submitted work.

Do you think you can waive these concerns? Or should I step down as a reviewer? Sorry that I only bring this up now ... the checklist reminded me.
