### 2022-08-11 #4595 @agricolab
I collaborate with @cbrnr on [LSL](https://github.com/sccn/labstreaminglayer) and [XDF](https://github.com/sccn/xdf), which might be considered a COI. Glad to see therefore that enough reviewers have already been found.

I am confident that MNELAB will find the reviewers approval. Looking forward to see it published.
