### 2022-08-14 #4595 @vahid-sb
@Kevin-Mattheus-Moerman the links to pdf generated by editorialbot are broken, and it does not generate the checklist.
