### 2022-08-14 #4595 @sappelhoff
@vahid-sb you have to post the command in this thread:

- https://github.com/openjournals/joss-reviews/issues/4650

see (emphasis by me):

> Checklists can only be created once the review has started **in the review issue**

(this is the **pre**-review issue)
