### 2022-08-30 cbrnr/mnelab#359 @vahid-sb
JOSS review - vahid-sb
Hi guys
As part of the review of your JOSS submissions, I have been testing MNELAB for the past few days. And decided to open this issue to mention anything that need fixing. I will add more if I notice more issues. But before that, I should congratulate you all on this repository. Obviously, there is still many features that can be added, but as it is, it is already a very impressive and useful library.
I tested MNELAB under both Linux (Ubuntu 21) and Windows 10. In both cases, I created a conda environment using
`conda create -n mnelab python=3.9`
and after activating the environment, installed mnelab using:
`pip install mnelab[full]`
In both cases, I used output data from Branvision Acticap-64 and also from Wearable Sensing DSI-7 recorders.

1. The first issue is that in Windows, I can plot the time-series data, but when trying to plot PSD, I get the following error:

```
Traceback (most recent call last):
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mnelab\mainwindow.py", line 791, in plot_psd
    fig = self.model.current["data"].plot_psd(show=False, **kwds)
  File "<decorator-gen-220>", line 12, in plot_psd
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mne\io\base.py", line 1548, in plot_psd
    return plot_raw_psd(self, fmin=fmin, fmax=fmax, tmin=tmin, tmax=tmax,
  File "<decorator-gen-187>", line 12, in plot_raw_psd
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mne\viz\raw.py", line 438, in plot_raw_psd
    fig = _psd_figure(
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mne\viz\_mpl_figure.py", line 2172, in _psd_figure
    psd, freqs = psd_func(inst, tmin=tmin, tmax=tmax, picks=picks,
  File "<decorator-gen-138>", line 12, in psd_welch
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mne\time_frequency\psd.py", line 287, in psd_welch
    return psd_array_welch(data, sfreq, fmin=fmin, fmax=fmax, n_fft=n_fft,
  File "<decorator-gen-137>", line 12, in psd_array_welch
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mne\time_frequency\psd.py", line 202, in psd_array_welch
    f_spect = parallel(my_spect_func(d, func=func, freq_sl=freq_sl,
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\joblib\parallel.py", line 1001, in __call__
    self._print("Using backend %s with %d concurrent workers.",
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\joblib\parallel.py", line 871, in _print
    writer = sys.stderr.write
AttributeError: 'NoneType' object has no attribute 'write'
```

While in Linux I could plot PSD with no error.
