### 2022-08-31 cbrnr/mnelab#359 @vahid-sb
2. When opening a file which is not in a standard EEG format, I get the following error. To make the error message more informative for the user, please add an exception handling to the below mentioned line with a proper message that states file type not supported, rather than "_TypeError: cannot unpack non-iterable NoneType object_".

Traceback (most recent call last):
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mnelab\mainwindow.py", line 578, in open_data
    self.model.load(fname)
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mnelab\model.py", line 36, in wrapper
    f(*args, **kwargs)
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mnelab\model.py", line 110, in load
    data = read_raw(fname, *args, **kwargs, preload=True)
  File "C:\Users\DNI-2\miniconda3\envs\mnelab2\lib\site-packages\mnelab\io\readers.py", line 79, in read_raw
    _, ext = split_name_ext(fname)
TypeError: cannot unpack non-iterable NoneType object
