### 2022-09-11 #4650 @tuliofalmeida
Hi @Kevin-Mattheus-Moerman, sorry for not doing this sooner. I recommend acceptance, the interface worked well in my tests and has good documentation. I understand that the features of the MNE will be implemented in the future, and they met what they proposed.
