### 2022-09-11 #4650 @vahid-sb
> @vahid-sb thank again for your help here. Can you pick up this review again. I see you several boxes unticked. Can you summarise to @cbrnr what actions you expect for these?

@Kevin-Mattheus-Moerman I have opened an Issue in the MNELAB repo two weeks ago, currently includes two issues needs resolving, and have been waiting for the authors to resolve those issues. When done I can continue with the review.
