### 2022-09-13 cbrnr/mnelab#359 @cbrnr
Regarding the first issue, you have hit an edge case which occurs only on Windows using the launcher `mnelab` and `joblib` installed. Among other things, this is because the `mnelab` launcher, which is auto-generated by `pip`, redirects stdout and stderr, which leads to this error.

There are three potential solutions:

1. Start MNELAB with `python -m mnelab` instead of `mnelab`.
2. Uninstall `joblib`, then everything works even with the `mnelab` launcher.
3. Update MNE-Python to the latest dev version (`pip install -U git+https://github.com/mne-tools/mne-python#egg=mne`); it seems like the recent changes related to computing PSDs also fixed this issue, so everything works even with the `mnelab` laucher.

MNE 1.2 will be released very soon, and then solution 3 will be automatically implemented. Therefore, could you please check if those three options (and especially the last one) work for you? If so, I think it's OK to wait until MNE 1.2 has been released.

I will comment on the second issue soon.
