+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Show some keyboard shortcuts when no file is open ([#350](https://github.com/cbrnr/mnelab/pull/350) by [Clemens Brunner](https://github.com/cbrnr))
 - Always show a notification when running the dev version ([#351](https://github.com/cbrnr/mnelab/pull/351) by [Clemens Brunner](https://github.com/cbrnr))
 
+### Fixed
+- Helpful error message when opening an unsupported file type ([#360](https://github.com/cbrnr/mnelab/pull/360) by [Clemens Brunner](https://github.com/cbrnr))
+
 ## [0.8.4] - 2022-05-05
 ### Added
 - Support importing 1D arrays from .mat files ([#348](https://github.com/cbrnr/mnelab/pull/348) by [Clemens Brunner](https://github.com/cbrnr))
