### 2022-09-13 f57068b
Improve unsupported file type error message (#360)
