### 2022-09-13 cbrnr/mnelab#359 @cbrnr
I've fixed the second issue in #360.
