### 2022-09-15 #4650 @cbrnr
Relevant issue is https://github.com/cbrnr/mnelab/issues/359.
