### 2022-09-18 cbrnr/mnelab#359 @vahid-sb
@cbrnr sorry for the delay in responding. I just found time to check both fixes and both issues are now resolved. Thanks for that. I will perform some more checks and tests and if I face no other issues, will update the review form accordingly.
