### 2022-10-06 #4650 @Kevin-Mattheus-Moerman
@cbrnr I have read you paper and it mostly looks like, I just have the following minor final comments:

- [ ] Check `...specific use cases:` I recommend just using `...specific use cases.`
- [ ] Check `... involve (1) specifying...` this should likely read `... involves (1) specifying...`
- [ ] Check `how these steps might look like` it is better to say `what these steps might look like`
- [ ] Check `which does not feature a GUI though`, although correct I would recommend `although it does not feature a GUI`.

Once you've looked at these, and if you've made changes, please call `@editorialbot generate pdf` to update the paper.
