### 2022-10-07 #4650 @cbrnr
Thanks @Kevin-Mattheus-Moerman, I applied your suggestions. I also added a figure, which I completely forgot in the first draft, but I think it's pretty important for a paper about a GUI. I hope that this is OK. The first page is a bit empty now, I tried moving some text, but it seems like I cannot change the figure position.
