### 2022-10-07 #4650 @Kevin-Mattheus-Moerman
@cbrnr I checked the changes and all looks good. I will now process this work for acceptance in JOSS. At this point can you work on the following:

- [ ] Please archive a copy of the reviewed software on [ZENODO](https://zenodo.org/) you can do this manually, or some find these automated steps useful: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
- [ ] Once the ZENODO archive is created please check the meta-data and amend it where needed. For instance the title should match the title of the paper, and that the author list should match as well (the list should only be for the paper authors, remove other contributors if they are listed). Furthermore you can add ORCID for authors if you like.
- [ ] Once the ZENODO archive is complete please report back here with the ZENODO DOI.
- [ ] Please can you inform me what the version tag for the published version should be? This review currently lists `v0.8.4`, but please let me know if that has changed in the mean time.
