+++ CONTRIBUTING.md
@@ -53,6 +53,6 @@ Follow these steps to make a new [PyPI](https://pypi.org/project/mnelab/) releas
 This concludes the new release. Now prepare the source for the next planned release as follows:
 
 - Update the `__version__` string to the next planned release and append `.dev0`
-- Start a new section at the top of `CHANGELOG.md` titled `[UNRELEASED] - XXXX-XX-XX`
+- Start a new section at the top of `CHANGELOG.md` titled `[UNRELEASED] - YYYY-MM-DD`
 
 Don't forget to push these changes!
