+++ .github/workflows/test.yml
@@ -14,7 +14,7 @@ jobs:
     - uses: actions/setup-python@v4
       with:
         python-version: '3.10'
-    - uses: pre-commit/action@v3.0.0
+    - uses: pre-commit/action@v3
 
   build-wheel:
     needs: style
