+++ .github/workflows/release.yml
@@ -14,7 +14,7 @@ jobs:
     - uses: actions/setup-python@v4
       with:
         python-version: '3.10'
-    - uses: pre-commit/action@v3.0.0
+    - uses: pre-commit/action@v3
 
   build-wheel:
     name: Build wheel
@@ -57,7 +57,7 @@ jobs:
         with:
           name: artifact
           path: dist
-      - uses: pypa/gh-action-pypi-publish@v1.5
+      - uses: pypa/gh-action-pypi-publish@v1
         with:
           user: __token__
           password: ${{ secrets.PYPI_API_TOKEN }}
