### 2022-10-08 9e57197
Use correct version tag for publish action
