### 2022-10-08 #4650 @cbrnr
> Please archive a copy of the reviewed software on [ZENODO](https://zenodo.org/) you can do this manually, or some find these automated steps useful: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content

Done. I messed up a bit, so there are now multiple v0.8.5 releases on Zenodo. But I guess that's OK, because the latest one works.

> Once the ZENODO archive is created please check the meta-data and amend it where needed. For instance the title should match the title of the paper, and that the author list should match as well (the list should only be for the paper authors, remove other contributors if they are listed). Furthermore you can add ORCID for authors if you like.

How do I do that? Adding a `.zenodo.json` means that I'd have to keep the version (and possibly other fields) up to date manually, which is something I'd like to avoid. What is the best practice here? Is creating a `CITATION.cff` sufficient, and can I skip the Zenodo DOI in this file (because it changes with every new release)?

> Please can you inform me what the version tag for the published version should be? This review currently lists v0.8.4, but please let me know if that has changed in the mean time.

I made a new v0.8.5 release, might depending on how I should handle the metadata update, yet another release might be necessary.
