### 2022-10-09 #4650 @Kevin-Mattheus-Moerman
@cbrnr editing the archive is not that difficult/technical. You can go to your ZENODO archive link and then simply click edit.

You can check all fields, but the main things to check and edit are:

- [ ] The title, which should match the paper
- [ ] The author list, which should match the paper
- [ ] The license, which should list your software license (leave as "open open", if I am not mistaken, if this does not work for you)
- [ ] Recommended: Adding ORCIDs for all authors

See here on how to get to the editing:
![Peek 2022-10-09 09-39](https://user-images.githubusercontent.com/8392709/194746857-34603281-2d5c-4230-9676-a6664d15d5ee.gif)
