### 2022-10-10 #4650 @Kevin-Mattheus-Moerman
@openjournals/dev can you help identify the problem? The authors formatted the bib entry as follows:

```bib
@article{Delorme:2004,
    author = {Delorme, A. and Makeig, S.},
    title = {{EEGLAB: an open source toolbox for analysis of single-trial EEG dynamics including independent component analysis}},
    journal = {Journal of Neuroscience Methods},
    volume = 134,
    number = 1,
    year = 2004,
    pages = 9--21,
    doi = {10.1016/j.jneumeth.2003.10.009}
}
```
Looks like the issue is with ` pages = 9--21`, should this perhaps be ` pages = "9--21"` instead?
