### 2022-10-10 #4650 @Kevin-Mattheus-Moerman
@cbrnr perhaps you can try the above in the meantime and call `@editorialbot check references` to see if it passes.
