### 2022-10-10 #4650 @cbrnr
Thanks, I added curly braces and it seems to work now!
